+++ example/lib/gallery/sections/icons.dart
@@ -460,6 +460,11 @@ class IconsSection extends StatelessWidget {
                         is16x16: true,
                       ),
                       _IconEntry(
+                        data: NesIcons16.leftArrowIndicator,
+                        label: 'leftArrowIndicator',
+                        is16x16: true,
+                      ),
+                      _IconEntry(
                         data: NesIcons16.bell,
                         label: 'bell',
                         is16x16: true,