+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - feat: adding `NesIcons.bell` and `NesIcons16.bell`
  - feat: adding `NesIcons16.close`
  - feat: adding `NesIcons16.handPointingRight`
+ - feat: adding `NesIcons16.leftArrowIndicator`
 
 # 0.18.0
  - feat: add `NesIcons.wrench`